Revise description for "delete project" command (#23)
This is a suggestion / request to re-word the delete project command and make it consistent with existing label "Remove project(s) `M-D'".  "Delete" can be a dangerous verb and the description doesn't entirely preclude the possibility that files could be deleted.  I suggest "remove" with a slightly more verbose explanation of the action.

What do you think?

Thanks for the package :-)  :+1: