Fix type for helm-projectile-sources-list variable
According to the
https://www.gnu.org/software/emacs/manual/html_node/elisp/Composite-Types.html
`list` is a composite type and should be used when the number of
elements and its types are known. For `helm-projectile-sources-list` the
better option would be `(repeat symbol)`.

emacs-29 introduces `setopt` macro to set customizable variables, it
verifies variable type according to `:type` annotation and throw an
error if type mismatches.