Deep copy list of known projects to avoid helm adding faces
I am not sure if this is the correct way to fix, yet this workaround seems to
work in my case. I had very similar observations as others claimed in #185 when
trying to debug this issue. Namely, right after executing
`helm-projectile-swithc-project` I observed that `projectile-known-projects`
had names changed with fontification.

Fixes #185