refactor!: Use built-in helm-grep-ag instead of obsolete helm-ag
The `helm-ag` has been removed from MELPA
https://github.com/melpa/melpa/pull/9520, with a recommendation to use a Helm
built-in `helm-grep-ag`.

This commit updates implementation of command `helm-projectile-ag` to use the
`helm-grep-ag`.

fixes: #193

BREAKING CHANGE: Default input is not populated (to be fixed in a following
  commit).

BREAKING CHANGE: Directories to keep (see `projectile-parse-dirconfig-file`)
  are no longer passed to searcher invocation.

BREAKING CHANGE: The Helm source that present search results is now
  `helm-grep-ag-class`, which results in how results are presented and
  interacted with. For example actions are now different, and "Edit search
  results" is no longer available. Type TAB (or C-i) in search results to see
  available actions.