build: Tune flycheck Elisp checkers for the project
Setting the variabl `flycheck-emacs-lisp-load-path` to `inherit` allows to load
packages installed for an Emacs session. This prevents `package-lint`
diagnostics about non-existing `require`s.

Update the `package-lint--sane-prefixes` to suppress `package-lint` diagnostics
about all variables that names start with `helm-source-`. The latter is a
recommended way of naming a source variable by Helm Wiki [1], which says:

> For consistency, prefix your source names with `helm-source-`
  (e.g. `helm-source-info-emacs`).

Alternative to the latter could be to rename all classes and variables that
names start with `helm-source-` and then, optionally - if backward
compatibility is required - to add `defalias`es for them with original names.

[1] https://github.com/emacs-helm/helm/wiki/Developing#creating-a-source