fix: Ensure helm-rg--extra-args are not changed in helm-projectile-rg
Remove side effect of updating a global value of `helm-rg--extra-args`.  Should
the value be changed it results in issues like changing
`helm-projectile-ignore-strategy` to `'search-tool` having `helm-projectile-rg`
still use the old value for `helm-rg--extra-args`.

This has been previously reported and already solved by @vigneshsarma in #191,
however, due to conflict with #192 the fix needed to be re-implemented.

closes: #191