feat: Add helm-projectile commands to switch to buffer in other window/frame
While technically the `helm-projectile-switch-to-buffer` allows to switch to
buffer in other window/frame (with actions bound to `C-c o` and `C-c C-o`
respectively), yet for some (like me) a second nature is to use an extra
`4` or `5` just before last keypress to switch a buffer. And indeed `projectile`
does provide `C-c p 4 b` and `C-c p 5 b` to switch to a project buffer in other
window/frame. Adding a substitute `helm-projectile` specific commands for that.