feat: Add helm-projectile commands to find file in other window/frame
While technically the `helm-projectile-find-file` allows to find a file other
window/frame (with actions bound to `C-c o` and `C-c C-o` respectively), yet
for some (like me) a second nature is to use an extra `4` or `5` just before
last keypress to find a file. And indeed `projectile` does provide `C-c p 4
f` and `C-c p 5 f` to find a project file in other window/frame. Adding
a substitute `helm-projectile` specific commands for that.