feat: Add helm-projectile commands to open directories in other window/frame
While technically the `helm-projectile-find-dir` allows to find directories in
other window/frame (with actions bound to `C-c o` and `C-c C-o` respectively),
yet for some (like me) a second nature is to use an extra `4` or `5` just
before last keypress to find a directory. And indeed `projectile` does provide
`C-c p 4 d` and `C-c p 5 d` to find a project directory in other window/frame
it will not be applied. Adding a substitute `helm-projectile` specific commands
for that.