feat: Add helm-projectile commands to find other file in other window/frame
While technically the `helm-projectile-other-find-file` allows to find an other
file in other window/frame (with actions bound to `C-c o` and `C-c C-o`
respectively), yet for some (like me) a second nature is to use an extra `4` or
`5` just before last keypress to find other file. And indeed `projectile` does
provide `C-c p 4 a` and `C-c p 5 a` to find a project other file in other
window/frame. Adding a substitute `helm-projectile` specific commands for that.