feat: Add helm-projectile commands to switch project in other window/frame
While technically the `helm-projectile-switch-project` allows to find files in
other window/frame (with actions bound to `C-c o` and `C-c C-o` respectively),
yet for some (like me) a second nature is to use an extra `4` or `5` just
before last keypress to switch to a project. While `projectile` doesn't provide
`C-c p 4 d` and `C-c p 5 d` to switch project and find a file in other
window/frame adding - in a hopefully future proof way - `helm-projectile`
specific commands for that.