feat: Add autoloads for commands defined with helm-projectile-command macro
While it's not necessary to have autoloads for these commands it makes user
life easier. For example, when defining their configuration with something
like the following:

(use-package helm-projectile
  :bind
  (("C-c H"   . #'helm-projectile-switch-project))
  :config
  (helm-projectile-on))

Should there be no autoload for the command `helm-projectile-switch-project`
byte compiler - and in consequence `flycheck` - complains that the command
might not be defined at runtime.