fix: Ensure searching in directory is performed after installing helm-rg
This was a bug in a new feature that allows to run search with rg on an
arbitrary project directory (see 8ec008ed6773f9d10812dd3dc543b5cb2903e976). The
but has only affected the new feature, as it won't search in the selected
project directory, but will search in the whole project instead.