feat!: Show current buffer in buffer list
Show current buffer when switching buffer.  The current buffer is pushed to the
end of the list though.

This is useful, when one wants to see different parts of the same buffer in
multiple different windows.

A new variable `helm-projectile-remove-current-buffer` has been added.  When
set to non nil, current buffer is not presented as a candidate when switching
source.

fixes: #177
closes: #175

BREAKING CHANGE: Current buffer is now displayed when calling
  `helm-projectile`, `helm-projectile-switch-to-buffer',
  `helm-projectile-switch-to-buffer-other-window', and
  `helm-projectile-switch-to-buffer-other-frame'.  Set
  `helm-projectile-remove-current-buffer` to non nil to restore previous
  behaviour (i.e., to omit current buffer from buffer list).